fix get long short data
